requirements.txt
fastapi==0.115.12
fastapi[standard]==0.115.12
starlette==0.46.2
uvicorn==0.34.2
qdrant-client==1.14.2
pydantic-settings==2.9.1
black==25.1.0
python-decouple==3.8
